+++ src/group/stabchain/builder/random/random_ift.rs
@@ -1,6 +1,8 @@
+use super::parameters::RandomAlgoParameters;
+
 use crate::group::orbit::abstraction::FactoredTransversalResolver;
-use crate::group::orbit::transversal::shallow_transversal::{
-    representative_raw_as_word, shallow_transversal,
+use crate::group::orbit::transversal::factored_transversal::{
+    factored_transversal_complete_opt, representative_raw_as_word,
 };
 use crate::group::stabchain::element_testing::residue_as_words_from_words;
 use crate::group::stabchain::{MovedPointSelector, Stabchain, StabchainRecord};
@@ -11,27 +13,20 @@ use crate::group::utils::{
 use crate::group::Group;
 use crate::perm::actions::SimpleApplication;
 use crate::perm::{Action, Permutation};
+use crate::DetHashMap;
 use itertools::Itertools;
 use rand::rngs::ThreadRng;
-use rand::{seq::IteratorRandom, Rng};
+use rand::seq::IteratorRandom;
+use rand::Rng;
 use std::cell::RefCell;
-use std::iter::{repeat_with, Iterator};
-
-//Constants for subproduct generation
-const C1: usize = 1;
-const C2: usize = 1;
-const ORBIT_BOUND: usize = 50;
-const BASE_BOUND: usize = 5;
-// Cosntants for the Strong Generating Test.
-const C3: usize = 1;
-const C4: usize = 1;
+use std::collections::VecDeque;
+use std::iter::{repeat_with, FromIterator};
 
 // Helper struct, used to build the stabilizer chain
 
-pub struct StabchainBuilderRandomSTrees<P, S, A = SimpleApplication<P>, R = ThreadRng>
+pub struct StabchainBuilderRandom<P, S, A = SimpleApplication<P>, R = ThreadRng>
 where
     A: Action<P, OrbitT = usize>,
-    R: rand::Rng,
 {
     current_pos: usize,
     chain: Vec<StabchainRecord<P, FactoredTransversalResolver<A>, A>>,
@@ -41,34 +36,33 @@ where
     n: usize,
     base: Vec<A::OrbitT>,
     rng: RefCell<R>,
+    constants: super::parameters::Constants,
     //The chain is zero indexed, but this field is 1 indexed.
     //This is due to the end condition being when the chain is up to date below the index of the first record position, and this would be -1 with zero indexing.
     //For zero indexing this would have to be a signed type, which doesn't really seem worth it just to require one negative value at the end condition.
     up_to_date: usize,
-    //Store the depths of each of the schrier trees.
-    depths: Vec<usize>,
-    original_generators: Group<P>,
 }
 
-impl<P, S, A, R> StabchainBuilderRandomSTrees<P, S, A, R>
+impl<P, S, A, R> StabchainBuilderRandom<P, S, A, R>
 where
     P: Permutation,
     S: MovedPointSelector<P, A::OrbitT>,
     A: Action<P, OrbitT = usize>,
-    R: Rng + Clone,
+    R: Rng,
 {
-    pub fn new(selector: S, action: A, random: R) -> Self {
-        StabchainBuilderRandomSTrees {
+    pub fn new(selector: S, action: A, params: RandomAlgoParameters<R>) -> Self {
+        let (constants, random) = params.consts();
+
+        StabchainBuilderRandom {
             current_pos: 0,
             chain: Vec::new(),
             selector,
             action,
             n: 0,
             base: Vec::new(),
+            constants,
             rng: RefCell::new(random),
             up_to_date: 1,
-            depths: vec![],
-            original_generators: Group::new(&[]),
         }
     }
 
@@ -105,17 +99,12 @@ where
         let moved_point = self.selector.moved_point(moved_point_generator);
         //Create the top level record for this chain, and add it to the chain.
         //TODO check if you should add generators 1 by 1, in case there are redundant generators.
-        let mut initial_gens = group.clone();
-        self.original_generators = initial_gens.clone();
-        let (transversal, initial_depth) = shallow_transversal(
-            &mut initial_gens,
+        let initial_record = StabchainRecord::new(
             moved_point,
-            &self.action,
-            &mut *self.rng.borrow_mut(),
+            group.clone(),
+            factored_transversal_complete_opt(&group, moved_point, &self.action),
         );
-        let initial_record = StabchainRecord::new(moved_point, initial_gens, transversal);
         self.base.push(moved_point);
-        self.depths.push(initial_depth);
         self.chain.push(initial_record);
         self.sgc();
     }
@@ -127,8 +116,13 @@ where
         coset_representatives: usize,
         gens: &[P],
     ) -> Vec<Vec<P>> {
-        // Sum of all the depths in the tree.
-        let t: usize = self.depths.iter().sum();
+        //Sum of all "depths". In reality the transversal doesn't have a depth, so we use this as a upper bound.
+        let t = self
+            .chain
+            .iter()
+            .map(|record| (record.transversal.len() as f64).log2())
+            .sum::<f64>()
+            .floor() as usize;
         let record = &self.chain[self.current_pos];
         let k = rand::Rng::gen_range(&mut *self.rng.borrow_mut(), 0, 1 + gens.len() / 2);
         //Create an iterator of subproducts w and w2
@@ -155,7 +149,6 @@ where
                         record.base,
                         *point,
                         &self.action,
-                        self.depths[self.current_pos],
                     )
                     .unwrap()
                 })
@@ -176,19 +169,51 @@ where
 
     /// Check if adding a new element modifies the current layer of the chain.
     fn check_transversal_augmentation(&mut self, p: P) {
-        let mut record = &mut self.chain[self.current_pos];
-        debug_assert!(!record.gens.generators.contains(&p));
-        record.gens.generators.push(p);
-        //Calculate a new shallow transversal.
-        let (transversal, new_depth) = shallow_transversal(
-            &mut record.gens,
-            record.base,
-            &self.action,
-            &mut *self.rng.borrow_mut(),
-        );
-        record.transversal = transversal;
-        //Update the depths of the current position.
-        self.depths[self.current_pos] = new_depth;
+        let mut record = self.chain[self.current_pos].clone();
+        //debug_assert!(record.gens.generators().contains(&p));
+        // If this element is already a generator, then we can exit
+        let mut to_check = VecDeque::from_iter(record.transversal.keys().cloned());
+        let mut new_transversal = DetHashMap::default();
+        while !to_check.is_empty() {
+            let orbit_element = to_check.pop_back().unwrap();
+            let new_image = self.action.apply(&p, orbit_element);
+
+            // If we haven't seen this element.
+            if !(record.transversal.contains_key(&new_image)
+                || new_transversal.contains_key(&new_image))
+            {
+                new_transversal.insert(new_image, p.inv());
+            }
+        }
+
+        // We now want to check all the newly added elements
+        let mut to_check = VecDeque::from_iter(new_transversal.keys().cloned());
+
+        // Update the record
+        record.transversal.extend(new_transversal);
+
+        // While we have orbit elements (and representatives to check)
+        while !to_check.is_empty() {
+            // Get the pair
+            let orbit_element = to_check.pop_back().unwrap();
+            // For each generator (and p)
+            for generator in std::iter::once(&p).chain(record.gens.generators()) {
+                let new_image = self.action.apply(&generator, orbit_element);
+                // If we haven't already seen the image
+                record.transversal.entry(new_image).or_insert_with(|| {
+                    // Update and ask to check the new image
+                    to_check.push_back(new_image);
+                    generator.inv()
+                });
+            }
+        }
+        // Update the generators adding p if it isn't already present.
+        if !record.gens.generators().contains(&p) {
+            record.gens =
+                Group::from_iter(std::iter::once(&p).chain(record.gens.generators()).cloned());
+        }
+        // Store the updated record in the chain
+        self.chain[self.current_pos] = record;
     }
 
     ///Check if the permutation augments the orbit at a level, resetting the position afterwards.
@@ -214,7 +239,8 @@ where
             .cloned()
             .collect::<Vec<P>>();
         //Random products of the form gw
-        let mut random_gens = self.random_schrier_generators_as_word(C1, C2, &gens[..]);
+        let mut random_gens =
+            self.random_schrier_generators_as_word(self.constants.c1, self.constants.c2, &gens[..]);
         //Convert these into random schrier generators, by concatenating the resdiue of the inverse to it.
         random_gens.iter_mut().for_each(|gw| {
             //Get the residue of this word
@@ -228,28 +254,29 @@ where
             let (drop_out_level, h_residue) = residue_as_words_from_words(self.current_chain(), &h);
             if self.sifted(drop_out_level) {
                 //Pick the points that should be evaluated. This is a heuristic to speed up run times.
-                let evaluated_points: Vec<A::OrbitT> = if record.transversal.len() <= ORBIT_BOUND {
-                    //Evaluate on all points of the current orbit.
-                    record.transversal.keys().cloned().collect()
-                } else if self.base.len() <= BASE_BOUND {
-                    //Evaluate on BASE_BOUND randomly chosen points.
-                    record
-                        .transversal
-                        .keys()
-                        .choose_multiple(&mut *self.rng.borrow_mut(), BASE_BOUND)
-                        .into_iter()
-                        .cloned()
-                        .collect()
-                } else {
-                    //Evaluate on b_star randomly chosen points.
-                    record
-                        .transversal
-                        .keys()
-                        .choose_multiple(&mut *self.rng.borrow_mut(), b_star.len())
-                        .into_iter()
-                        .cloned()
-                        .collect()
-                };
+                let evaluated_points: Vec<A::OrbitT> =
+                    if record.transversal.len() <= self.constants.orbit_bound {
+                        //Evaluate on all points of the current orbit.
+                        record.transversal.keys().cloned().collect()
+                    } else if self.base.len() <= self.constants.base_bound {
+                        //Evaluate on BASE_BOUND randomly chosen points.
+                        record
+                            .transversal
+                            .keys()
+                            .choose_multiple(&mut *self.rng.borrow_mut(), self.constants.base_bound)
+                            .into_iter()
+                            .cloned()
+                            .collect()
+                    } else {
+                        //Evaluate on b_star randomly chosen points.
+                        record
+                            .transversal
+                            .keys()
+                            .choose_multiple(&mut *self.rng.borrow_mut(), b_star.len())
+                            .into_iter()
+                            .cloned()
+                            .collect()
+                    };
                 //If any point is not fixed by the residue, then we add the residue as a generator.
                 if !self.is_trivial_residue(&h_residue, evaluated_points) {
                     //Not all permutations have been discarded
@@ -261,15 +288,10 @@ where
                     debug_assert!(!self.base.contains(&new_base_point));
                     self.base.push(new_base_point);
                     //Fields for the new record.
-                    let mut gens = Group::new(&[h_star]);
-                    let (transversal, depths) = shallow_transversal(
-                        &mut gens,
-                        new_base_point,
-                        &self.action,
-                        &mut *self.rng.borrow_mut(),
-                    );
+                    let gens = Group::new(&[h_star]);
+                    let transversal =
+                        factored_transversal_complete_opt(&gens, new_base_point, &self.action);
                     let record = StabchainRecord::new(new_base_point, gens, transversal);
-                    self.depths.push(depths);
                     self.chain.push(record);
                     //Now up to date beneath the newly added point.
                     self.up_to_date = self.base.len() + 1;
@@ -311,12 +333,16 @@ where
             .cloned()
             .collect::<Vec<P>>();
         //Create an iterator that first has the original generators, and then the random schrier generators.
-        let products: Vec<Vec<P>> = self
-            .original_generators
-            .generators()
+        let products: Vec<Vec<P>> = self.chain[0]
+            .gens
+            .generators
             .iter()
             .map(|p| vec![p.clone()])
-            .chain(self.random_schrier_generators_as_word(C3, C4, &gens[..]))
+            .chain(self.random_schrier_generators_as_word(
+                self.constants.c3,
+                self.constants.c4,
+                &gens[..],
+            ))
             .collect();
         //Sift the original generators, and all products of the form g*w_{1,2}.
         for p in products {
@@ -338,18 +364,12 @@ where
             let collapsed_residue = collapse_perm_word(&residue);
             //If this point sifted through but isn't trivial, then we need a new record and base point.
             if self.sifted(drop_out_level) {
-                //TODO add function to add a new level
                 let moved_point = self.selector.moved_point(&collapsed_residue);
-                let mut gens = Group::new(&[collapsed_residue]);
-                let (transversal, depth) = shallow_transversal(
-                    &mut gens,
-                    moved_point,
-                    &self.action,
-                    &mut *self.rng.borrow_mut(),
-                );
+                let gens = Group::new(&[collapsed_residue]);
+                let transversal =
+                    factored_transversal_complete_opt(&gens, moved_point, &self.action);
                 let initial_record = StabchainRecord::new(moved_point, gens, transversal);
                 self.base.push(moved_point);
-                self.depths.push(depth);
                 self.chain.push(initial_record);
                 self.up_to_date = self.base.len() + 1;
             } else {
@@ -386,13 +406,13 @@ where
     }
 }
 
-impl<P, S, A, R> super::Builder<P, FactoredTransversalResolver<A>, A>
-    for StabchainBuilderRandomSTrees<P, S, A, R>
+impl<P, S, A, R> crate::group::stabchain::builder::Builder<P, FactoredTransversalResolver<A>, A>
+    for StabchainBuilderRandom<P, S, A, R>
 where
     P: Permutation,
     A: Action<P, OrbitT = usize>,
     S: MovedPointSelector<P, A::OrbitT>,
-    R: Rng + Clone,
+    R: Rng,
 {
     fn set_generators(&mut self, gens: &Group<P>) {
         self.construct_strong_generating_set(gens);
